CSS: replace checkbox with toggle switch
Researched around and custom styling of checkbox is beyond savaging on
Firefox.  The only way to make this work is to have everything custom,
which is non-trivial.  Replaced it with a toggle switch instead because
of the simplicity and since it's an immediate on-off switch as well
which is more fitting to use a toggle switch for.